algo_trader/requirements.txt
alpaca-py>=0.30.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b1
loguru>=0.7.0
python-dateutil>=2.8.0
pytz>=2023.3
aiohttp>=3.9.0
schedule>=1.2.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
